pyproject.toml
[project]
name = "antigravity-3tier-crewai"
version = "2.0.0"
description = "3-Tier Multi-Agent Architecture with CrewAI Integration"
requires-python = ">=3.11"
dependencies = [
    "crewai[openai,litellm]>=0.80.0",
    "crewai-tools>=0.75.0,<0.76.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "ruamel.yaml>=0.17.32",
    "litellm>=1.15.0,<1.74",
    "tokenizers>=0.21.0",
    "tiktoken>=0.9.0",
    "pydantic>=2.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0"
]
experimental = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "pydantic-ai>=0.0.13"
]

[dependency-groups]
dev = ["jupyter"]

[tool.black]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git", ".venv", "node_modules", "__pycache__", ".pytest_cache"]
